+++ src/torchjd/aggregation/_krum.py
@@ -10,7 +10,7 @@ from ._weighting_bases import GramianWeighting
 
 class KrumWeighting(GramianWeighting):
     """
-    :class:`~torchjd.aggregation._weighting_bases.Weighting` giving the weights of
+    :class:`~torchjd.aggregation.GramianWeighting` giving the weights of
     :class:`~torchjd.aggregation.Krum`.
 
     :param n_byzantine: The number of rows of the input matrix that can come from an adversarial
@@ -81,7 +81,7 @@ class KrumWeighting(GramianWeighting):
 
 class Krum(GramianWeightedAggregator):
     """
-    :class:`~torchjd.aggregation._aggregator_bases.Aggregator` for adversarial federated learning,
+    :class:`~torchjd.aggregation.GramianWeightedAggregator` for adversarial federated learning,
     as defined in `Machine Learning with Adversaries: Byzantine Tolerant Gradient Descent
     <https://proceedings.neurips.cc/paper/2017/file/f4b9ec30ad9f68f89b29639786cb62ef-Paper.pdf>`_.