+++ src/torchjd/aggregation/_sum.py
@@ -7,7 +7,7 @@ from ._weighting_bases import MatrixWeighting
 
 class SumWeighting(MatrixWeighting):
     r"""
-    :class:`~torchjd.aggregation._weighting_bases.Weighting` that gives the weights
+    :class:`~torchjd.aggregation.MatrixWeighting` that gives the weights
     :math:`\begin{bmatrix} 1 & \dots & 1 \end{bmatrix}^T \in \mathbb{R}^m`.
     """
 
@@ -20,7 +20,7 @@ class SumWeighting(MatrixWeighting):
 
 class Sum(WeightedAggregator):
     """
-    :class:`~torchjd.aggregation._aggregator_bases.Aggregator` that sums of the rows of the input
+    :class:`~torchjd.aggregation.WeightedAggregator` that sums of the rows of the input
     matrices.
     """