+++ src/torchjd/aggregation/_trimmed_mean.py
@@ -6,7 +6,7 @@ from ._aggregator_bases import Aggregator
 
 class TrimmedMean(Aggregator):
     """
-    :class:`~torchjd.aggregation._aggregator_bases.Aggregator` for adversarial federated learning,
+    :class:`~torchjd.aggregation.Aggregator` for adversarial federated learning,
     that trims the most extreme values of the input matrix, before averaging its rows, as defined in
     `Byzantine-Robust Distributed Learning: Towards Optimal Statistical Rates
     <https://proceedings.mlr.press/v80/yin18a/yin18a.pdf>`_.